Add support for UICollectionView with SwipeCollectionViewCell
This is due in large part thanks to gcox and mschonvogel for the
initial implementation. This makes a copy of SwipeTableViewCell and
bases it on UICollectionViewCell, making plenty of necessary adjustments
to calculate translations and frames/centers based on the collection
view cell's `contentView` instead of just the cell itself (like the
table view cell implementation does). This means we can drag the cell
left and right without it overlapping neighbor cells (clipsToBounds is
true). Additionally, trigger points for the expansion animations are
calculated based on the cell's position within the collection view, and
not the whole collection view width like table views are.

As gcox mentioned in his original pull request,
`SwipeTableOptions` has been renamed to `SwipeOptions` since it's not
specific to tables in any way. There's also still a fair amount of
duplicated code between the two implementations, and it may be worth it
to refactor this to reduce that where it makes sense.

Additionally, the example app has been updated with a `UICollectionView`
example that uses a `UICollectionViewFlowLayout` set so that cells sit
side by side in a 2-wide grid.